mage: refactor compress
