+++ magefile.go
@@ -21,11 +21,8 @@
 package main
 
 import (
-	"archive/tar"
-	"compress/gzip"
 	"context"
 	"fmt"
-	"io"
 	"log"
 	"os"
 	"path/filepath"
@@ -37,6 +34,7 @@ import (
 
 	"github.com/elastic/beats/v7/dev-tools/mage"
 
+	"github.com/elastic/apm-server/beater/compress"
 	"github.com/elastic/apm-server/beater/config"
 )
 
@@ -259,82 +257,6 @@ func customizePackaging() {
 	}
 }
 
-func Tar(source, target string) error {
-	filename := filepath.Base(source)
-	target = filepath.Join(target, fmt.Sprintf("%s.tar", filename))
-	tarfile, err := os.Create(target)
-	if err != nil {
-		return err
-	}
-	defer tarfile.Close()
-
-	tarball := tar.NewWriter(tarfile)
-	defer tarball.Close()
-
-	info, err := os.Stat(source)
-	if err != nil {
-		return nil
-	}
-
-	var baseDir string
-	if info.IsDir() {
-		baseDir = filepath.Base(source)
-	}
-
-	return filepath.Walk(source,
-		func(path string, info os.FileInfo, err error) error {
-			if err != nil {
-				return err
-			}
-			header, err := tar.FileInfoHeader(info, info.Name())
-			if err != nil {
-				return err
-			}
-
-			if baseDir != "" {
-				header.Name = filepath.Join(baseDir, strings.TrimPrefix(path, source))
-			}
-
-			if err := tarball.WriteHeader(header); err != nil {
-				return err
-			}
-
-			if info.IsDir() {
-				return nil
-			}
-
-			file, err := os.Open(path)
-			if err != nil {
-				return err
-			}
-			defer file.Close()
-			_, err = io.Copy(tarball, file)
-			return err
-		})
-}
-
-func Gzip(source, target string) error {
-	reader, err := os.Open(source)
-	if err != nil {
-		return err
-	}
-
-	filename := filepath.Base(source)
-	target = filepath.Join(target, fmt.Sprintf("%s.gz", filename))
-	writer, err := os.Create(target)
-	if err != nil {
-		return err
-	}
-	defer writer.Close()
-
-	archiver := gzip.NewWriter(writer)
-	archiver.Name = filename
-	defer archiver.Close()
-
-	_, err = io.Copy(archiver, reader)
-	return err
-}
-
 func saveIronbank() error {
 	ironbank := getIronbankContextName()
 	buildDir := filepath.Join("build", ironbank)
@@ -350,19 +272,10 @@ func saveIronbank() error {
 		}
 	}
 
-	// Save the build context as artifact
-	Tar(buildDir, distributionsDir)
-	source := filepath.Join(distributionsDir, ironbank)
-	tarSource := source + ".tar"
-	Gzip(tarSource, distributionsDir)
-	tarGzSource := tarSource + ".gz"
-
-	// Remove leftovers
-	if err := os.RemoveAll(source); err != nil {
-		return errors.Wrapf(err, "failed to clean existing build directory %s", source)
-	}
-	if err := os.RemoveAll(tarSource); err != nil {
-		return errors.Wrapf(err, "failed to clean existing build directory %s", tarSource)
+	// Save the build context as tar.gz artifact
+	tarGzSource, err := compress.TarGz(buildDir, distributionsDir)
+	if _, err := os.Stat(buildDir); os.IsNotExist(err) {
+		return fmt.Errorf("cannot compress the tar.gz file")
 	}
 
 	return errors.Wrap(mage.CreateSHA512File(tarGzSource), "failed to create .sha512 file")
